% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conv.plot}
\alias{conv.plot}
\title{Create a plot to examine the convergence of the algorithm.}
\usage{
conv.plot(result, true_coeffs = NULL, x = NULL, y = NULL, family = binomial())
}
\arguments{
\item{result}{The results from irls.sgd.}

\item{true_coeffs}{A vector of true coefficients if they are available.}

\item{x}{A matrix of independent variables to estimate the true coefficients from.}

\item{y}{A vector with the dependent variable to estimate the true coefficients from.}

\item{family}{A glm family for the distribution to use, i.e. "binomial()" when estimating the true coefficients.}
}
\description{
Create a plot to examine the convergence of the algorithm.
}
