% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussianG}
\alias{gaussianG}
\title{Gaussian gradient}
\usage{
gaussianG(theta, data, idx)
}
\arguments{
\item{theta}{The coefficient vector}

\item{data}{A matrix with the data to use}

\item{idx}{The rows to use (base 0 for now)}
}
\description{
Gaussian gradient
}
