% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_deviance}
\alias{get_deviance}
\title{Get the deviance for a model}
\usage{
get_deviance(beta, x, y, family)
}
\arguments{
\item{beta}{The vector of coefficients}

\item{y}{The dependent variable}

\item{family}{A glm family for the distribution to use, i.e. "binomial()"}

\item{X}{A matrix containing the covariates (including an intercept if one wants to use one)}
}
\description{
Get the deviance for a model
}
