% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd.R
\name{glm.grad}
\alias{glm.grad}
\title{Gradient for GLM}
\usage{
glm.grad(x, y, theta, idx, family)
}
\arguments{
\item{x}{A matrix containing the covariates (including an intercept if one wants to use one).}

\item{y}{The dependent variable.}

\item{theta}{The coefficient vector to use.}

\item{family}{A glm family for the distribution to use, i.e. "binomial()".}
}
\description{
Gradient for GLM
}
