% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd.R
\name{glm.sgd}
\alias{glm.sgd}
\title{(Batch) (Stochastic) Gradient Descent for GLM}
\usage{
glm.sgd(x, y, family, sgd.ctrl = NULL)
}
\arguments{
\item{x}{A matrix containing the covariates (including an intercept if one wants to use one).}

\item{y}{The dependent variable.}

\item{family}{A glm family for the distribution to use, i.e. "binomial()".}

\item{sgd.ctrl}{control parameters to pass to the sgd function (see ?sgd for details).}
}
\description{
(Batch) (Stochastic) Gradient Descent for GLM
}
