% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irls.R
\name{irls}
\alias{irls}
\title{Subsampling IRLS}
\usage{
irls(
  X,
  y,
  family,
  ctrl = list(subs = 1, maxit = 100, tol = 1e-07, cooling = c(3, 0.9, 0.95), expl =
    c(3, 1.5))
)
}
\arguments{
\item{X}{A matrix containing the covariates (including an intercept if one wants to use one)}

\item{y}{The dependent variable}

\item{family}{A glm family for the distribution to use, i.e. "binomial()"}

\item{ctrl}{A list of control parameters}
}
\description{
Subsampling IRLS
}
\details{
The control parameters that can be set are
\itemize{
 \item{"subs"}{The subsample proportion to use for each iteration}
 \item{"maxit"}{The maximum number of iterations to run for}
 \item{"tol"}{The relative tolerance to consider when measuring convergence (not really useful for subsampling)}
 \item{"cooling"}{The cooling schedule to use, 3 numericals. First is number of iterations to stay constant for, second is the temperature to start the first cooling iteration at and third is the exponential decay.}
 \item{"expl"}{The explosion detection settings, 2 numericals. First is number of iterations to not detect at and the second is the relative change in deviance to consider an explosion.}
}
}
