\name{irls.sgd-package}
\alias{irls.sgd-package}
\docType{package}
\title{
\packageTitle{irls.sgd}
}
\description{
\packageDescription{irls.sgd}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{irls.sgd}
\packageIndices{irls.sgd}
Functions of interest are `irls`, `irls.sgd`, `sgd` and `glm.sgd`.
}
\author{
\packageAuthor{irls.sgd}

Maintainer: \packageMaintainer{irls.sgd}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
