% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irls.sgd.R
\name{irls.sgd}
\alias{irls.sgd}
\title{Subsampling IRLS with SGD}
\usage{
irls.sgd(
  x,
  y,
  family,
  irls.control = list(subs = 1, maxit = 100, tol = 1e-07, cooling = c(3, 0.9, 0.95),
    expl = c(3, 1.5)),
  sgd.control = list(subs = 1, alpha = 5e-04, ),
  save_hist = F
)
}
\arguments{
\item{x}{A matrix containing the covariates (including an intercept if one wants to use one)}

\item{y}{The dependent variable}

\item{family}{A glm family for the distribution to use, i.e. "binomial()"}

\item{irls.control}{A list of control parameters for the Subsampling IRLS (see ?irls for details).}

\item{sgd.control}{A list of control parameters for the SGD (see ?sgd for details, note that sgd.control$start will be overwritten).}
}
\description{
Subsampling IRLS with SGD
}
