% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd.R
\name{sgd}
\alias{sgd}
\title{(Batch) (Stochastic) Gradient Descent}
\usage{
sgd(
  gradient,
  data = NULL,
  ctrl = list(start, alpha, decay, subs, maxit, histfreq)
)
}
\arguments{
\item{gradient}{The gradient of the objective function.}

\item{data}{The dataset to use for the gradient, if applicable, if NULL, no data is used.}

\item{ctrl}{A list of control parameters.}
}
\description{
(Batch) (Stochastic) Gradient Descent
}
\details{
The control parameters that can be set are
\itemize{
 \item{"start" }{The starting point, NULL means random around 0.}
 \item{"alpha" }{The initial step size, defaults to 0.0005.}
 \item{"decay" }{The exponential decay of the step size per iteration, defaults to 0.999.}
 \item{"subs" }{The subsample percentage to use at each iteration, defaults to 1 (100\%).}
 \item{"maxit" }{The maximum number of iterations to run for, defaults to 1000.}
 \item{"histfreq" }{The number of iterations between stored history frames, defaults to 50.}
}
}
