% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgd.R
\name{sgd}
\alias{sgd}
\title{(Batch) (Stochastic) Gradient Descent}
\usage{
sgd(
  gradient,
  start = NULL,
  data = NULL,
  alpha = 5e-04,
  decay = NULL,
  subs = 1,
  maxit = 5000
)
}
\arguments{
\item{gradient}{The gradient of the objective function}

\item{start}{The starting point, NULL means random around 0}

\item{data}{The dataset to use for the gradient, if necessary}

\item{alpha}{The initial step size}

\item{decay}{The exponential decay of the step size per iteration}

\item{subs}{The subsample percentage to use at each iteration}

\item{maxit}{The maximum number of iterations to run for}
}
\description{
(Batch) (Stochastic) Gradient Descent
}
\details{
The control parameters that can be set are
}
